+++ obfutil/core/__init__.py
@@ -1,0 +1,3 @@
+from obfutil.core.api import ObfUtilAPI, api
+
+__all__ = ['ObfUtilAPI', 'api']