+++ obfutil/obfuscation/__init__.py
@@ -1,0 +1,3 @@
+from obfutil.obfuscation.core import obfuscate_code, NuclearObfuscator
+
+__all__ = ['obfuscate_code', 'NuclearObfuscator']